Make builds work when BUILD_DIR=/app
When building with either 'git push' or Anvil, BUILD_DIR is set to a
temporary directory (not /app). However, because the prebuilt Python
interpreter has a prefix of /app/.heroku/python baked into it, the
contents of BUILD_DIR need to be temporarily swapped into /app for the
rest of the buildpack to work.

This temporary swapping breaks other buildpack-compatible tools that
set BUILD_DIR=/app. So the fix is straightforward---skip the swapping
in this case.

Also alias bpwatch to a no-op when REQUEST_ID is not set, since this
indicates we're running outside the Heroku environment.