fix(utils): fix deep-rm searching behavior
When a child directory exists in a specified directory, the buildpack
would crash and burn. For example, with a sample Django app, you could see
this error:

```
rm: cannot remove `/app/notes': Directory not empty
```

This only occurs with projects external to Heroku (i.e.
https://github.com/flynn/slugbuilder)

This new behavior is symantically correct except that it will now
properly remove child directories.