Add a utility function for downloading & unpacking archives
Once we start verifying checksums, the amount of boilerplate is going
to increase, so it makes sense to refactor. The new function separates
out download and extraction, since we'll want to verify the checksum
between the two steps.

In addition, the curl download has been made more resilient against
failures - it now explicitly fails for non-HTTP 200 responses & also
retries. (Inspired by the implementation in heroku-buildpack-nodejs.)