Verify the checksums of packages downloaded from AWS
With pip 8, the integrity of packages downloaded from PyPI during the
buildpack compile can be ensured by adding hashes to the requirements
files. However until now there was no guarantee that the python and/or
library packages downloaded from AWS were had not been tampered with
themselves.

Whilst the downloads are performed over HTTPS, also verifying the hash
of the downloaded archives protects against:
* S3 bucket compromise
* CA miss-issuance of an s3.amazonaws.com cert
* Protocol or client exploits (Qualys site rating for S3 is only a "B")
* Non-malicious modification of existing files (eg uploading a newer
  Python version over an older one)

In the case of checksum mismatches, the compile will now be aborted.

For now, if a known checksum is not found for the downloaded file, then
only a warning is output. This will prevent breakage if there are any
runtimes or stacks uploaded to S3 that are not mentioned in this repo.
In the future we may want to make this cause the compile to fail.

The existing checksums were generated via a script run on a cedar-14
dyno (see the PR).