Update puts-step "Installing pip"
Since the output of the actual pip install is redirected into /dev/null, one never knows (by just reading the deploy output) which pip version is installed. Adding the version to the puts-step should make for quicker debugging. It could be considered to also echo the installed version in case it doesn't need to be installed (see line 62)