Remove unused pip-uninstall step
The pip-uninstall step has been unused for pip-using apps since #925,
since the buildpack now invalidates the entire package cache instead.

Whilst the step appears to still be used for pipenv-using apps, the
code is not run, since `SKIP_PIP_INSTALL=1` is set too early.

This bug was introduced in a334672a1a2814509f4bad06139786e497adbf3f
which landed straight to `master` two days after the feature was
introduced in #650.

Longer term we should likely get pipenv installs to do something similar
to pip (invalidate the whole cache based on checksum of the lockfile),
however for now I'm removing this deadcode since it's the last consumer
of the `pip-diff` script which we want to remove.

Closes @W-8386830@.