Add a make target for running a local compile
It's often useful to be able to run a compile outside of the unit or
Hatchet tests when developing the buildpack or debugging.

Whilst running the compile locally inside Docker won't fully replicate
a compile on the platform, for many use-cases it's close enough and
importantly gives a faster development feedback loop, since buildpack
changes don't have to be committed/pushed prior to triggering a compile.

Closes @W-8436406@.