Remove testing & binary generation support for Heroku-16
The Heroku-16 stack reached end-of-life on May 1st, 2021, and from
June 1st, 2021, builds will be disabled.

This removes support for testing against Heroku-16 in CI, and also
for building new Python binaries for the stack.

This change is non-breaking as it only affects development (hence why
it's landing a few days early), however a changelog entry has been added
regardless, for improved visibility into the support status of the stack.

Closes GUS-W-9329579.