Tests: Add additional coverage of editable package installation
Previously the editable installation mode test only checked whether
the build completed successfully, not whether the installed editable
packages worked.

Now the packages are tested via their entrypoints in the following scenarios:
1. During the `bin/post_compile` hook
2. By a later buildpack in the build
3. At runtime

In particular, (2) catches the issue described in #1006 when setuptools is upgraded,
and (1) + (3) will be useful to verify the solution to #1006 hasn't broken the other cases.

GUS-W-10047026.

[skip-changelog]