Remove build scripts for old CPython versions
Removes the build scripts for all CPython versions apart from the latest supported version of each major Python version, since we won't be
building these for future stacks, so have no need to keep the scripts
around (given they can be easily recreated in a DR scenario).

The already built archives on S3 are not affected.