Skip check for requirement.txt changes when using Pipenv (#1350)
Previously Pipenv builds would include a stray error message in the
log output, that didn't affect the result of the build, but added noise
to the logs and gave the false impression the build was broken in
some way:

```
cp: cannot stat '/tmp/build_a123c0456/requirements.txt': No such file or directory
```

This originated from the check of `requirements.txt` changes compared
to the copy of the file seen in the last build. This check is only necessary
when using Pip (and not Pipenv), so should always have been skipped.

Now that the check is fixed, the disabling of bash exit on error mode for
part of `bin/steps/python` can more easily be removed, increasing
coverage of such issues in the future.

Fixes #1270.
GUS-W-7924941.