Add warnings for EOL/deprecated Python versions (#1356)
Python 3.4, 3.5 and 3.6 have all reached end-of-life upstream,
so we should display a warning when they are used:
https://devguide.python.org/versions/#supported-versions

The existing EOL warning for Python 2.7 has been improved.

Lastly, a deprecation warning is now shown for builds using
the beta/unmaintained PyPy support. The PyPy versions
available on Heroku are outdated and insecure, and PyPy
usage represents less than 0.01% of builds on the platform.

GUS-W-8813901.
GUS-W-11603084.
GUS-W-11605715.