Provision a Heroku Postgres DB in fewer cases
As part of the Buildpack API, a buildpack can request the provisioning
of addons on the first build of an app via the `bin/release` payload.
(This feature does not apply to subsequent builds of existing apps.)

Previously the Python buildpack requested a Heroku Postgres DB be
provisioned if the app source contained a `manage.py` file in the
source root or any subdirectory up to three levels deep.

This meant a Postgres DB was provisioned even in cases where the
app wasn't a Django app (but just happened to have a `manage.py`
script) or used an alternative DB backend (eg MySQL), or didn't use
a DB at all.

Now, the DB is only provisioned if all of the following are true:
- a `manage.py` file exists in the root of the app source
- `django` is installed
- `psycopg2` is installed

This ensures it's still provisioned in the Python getting-started guide
case (and any other similar cases), but not in the other scenarios
mentioned above.

Fixes #1067.
GUS-W-8098791.