Fix automatic provisioning of Postgres DB addons
In #1363 the logic for determining when to automatically provision
a Heroku Postgres database addon was adjusted to reduce false
positives.

Unfortunately, this broke provisioning in all cases, since it turns
out that the build system does not source the `export` script prior
to running `bin/release` (like it does when running subsequent
buildpacks), and so the buildpack configured env vars (such as `PATH`
are not set.

This meant that when `bin/release` used the `is_module_available`
utility, it was calling system Python (where the Django and psycopg2
packages were not installed) and so always failing the conditional.

Unfortunately we do not currently have an easy way to test `bin/release`,
otherwise a test case would have been added in the original PR.

I had manually tested the original PR using the Python getting started
guide, but had only checked via accessing the app and trying to run
the Django `./manage.py migrate` command. However it seems the
Django config for that project silently falls back to sqlite if `DATABASE_URL`
is not set - so the project still worked.

For this PR I have manually tested to ensure the addon appears on the app.

 GUS-W-11851647.