Use the bundled pip wheel to perform pip bootstrap
Previously the buildpack downloaded a pip wheel from S3 in order
to bootstrap the pip/setuptools/wheel installation.

Newer Python versions (Python 3.4+) bundle a pip wheel as part of
the `ensurepip` module, which we can use instead of having to
perform this download now that we've dropped support for Python 2.

This approach:
- improves performance/reliability of repeat builds, since it saves
  having to download the pip wheel each time
- reduces the toil of performing pip updates, since we no longer need
  to upload new pip versions to the S3 bucket each time
- matches the approach already used in the new Python CNB.

GUS-W-13111316.