Remove support for Heroku-18
Since the Heroku-18 stack has reached end-of-life, and as such build using it
are no longer supported by the Heroku build system:
https://devcenter.heroku.com/changelog-items/2583

This fixes the integration tests failing in CI for the Heroku-18 stack, due to the
build system now (as expected) rejecting the jobs, eg:
https://github.com/heroku/heroku-buildpack-python/actions/runs/4852163562/jobs/8646809603?pr=1448#step:5:165

The Heroku-18 stack was the last to have Python binaries available for Python 2.7,
3.5 and 3.5, so the EOL conditionals and testcases for those versions have now been
removed.

Any non-Heroku consumers of this buildpack that wish to continue using
the Heroku-18 stack should pin to the previous version of this buildpack.

GUS-W-10446293.