Update AWS_DEFAULT_REGION to match the current S3 region
The AWS CLI requires that the `AWS_DEFAULT_REGION` env var
be set in order for authentication to work.

Previously `AWS_DEFAULT_REGION` was set to `us-west-2`, however,
this is different from the region our S3 buckets are in (`us-east-1`).

Whilst this doesn't affect the command we run at all, it caused confusion
recently as to what region the Python bucket was in, so for clarity the
default value has now been changed to match the region of the S3 bucket
`heroku-buildpack-python`.