Allow packaging tool versions to be updated by Dependabot
Previously the versions of the pip, setuptools, wheel and pipenv packaging
tools were stored as constants in the buildpack scripts. Now, they are stored
in requirements files, so that Dependabot can recognise and update them.

The PRs opened by Dependabot will still need human intervention (in the
form of adding appropriate changelog entries), however, this change will
at least save us from having to keep an eye out for new releases, and
reduces the amount of toil required compared to opening a PR from
scratch.

This is the classic buildpack equivalent of:
https://github.com/heroku/buildpacks-python/pull/29

GUS-W-13767545.