Fix the "security update available" version check
Previously the "Python security update is available" message was shown
if the requested Python version does not exactly equal what the
buildpack believes to be the latest Python patch version for that major
release.

However, this means that the message is then incorrectly shown if the
current version is actually newer than the latest version the buildpack
knows about.

This scenario can occur:
(a) During the small time window when a new Python version has been
    built and uploaded to S3 but the buildpack updates have not yet
    been released.
(b) If an app pins to an older buildpack version and then later
    updates to a new Python patch release.

It's not possible to add an integration test for this specific case, since
using a fake future version (like Python 3.12.999) will fail prior to the
version check due to it not existing on S3. (The scenario being fixed is
effectively a race condition that we can't emulate.)

However, the security version numbers are tested in general:
https://github.com/heroku/heroku-buildpack-python/blob/main/spec/hatchet/python_update_warning_spec.rb

GUS-W-15541305.